move project bbox to ProjectSerializer
